#77 refactor: 댓글 삭제 모달 별도 파일로 분리
